move encrypt/decrypt file functions to crypto.c
